Refactor and unskip clasp logs test
