+++ tests/constants.ts
@@ -43,9 +43,14 @@ const INVALID_CLASP_SETTINGS = {
   projectId: `project-id-${rndStr()}`,
 };
 
+const VALID_CLASP_SETTINGS_WITHOUT_PROJECT_ID = {
+  scriptId: SCRIPT_ID,
+};
+
 export const CLASP_SETTINGS = {
   valid: JSON.stringify(VALID_CLASP_SETTINGS),
   invalid: JSON.stringify(INVALID_CLASP_SETTINGS),
+  validWithoutProjectId: JSON.stringify(VALID_CLASP_SETTINGS_WITHOUT_PROJECT_ID),
 };
 
 const FAKE_CLASPRC_TOKEN = {
